+++ src/But/Log/ProxyThrowing.hpp
@@ -3,6 +3,7 @@
 #include <But/Format/apply.hpp>
 #include <But/Format/format.hpp>
 #include <But/Log/Destination/Sink.hpp>
+#include <But/Log/Destination/Common/rootElementTag.hpp>
 #include <But/Log/Localization/None.hpp>
 
 namespace But
@@ -28,20 +29,39 @@ public:
   ProxyThrowing(Destination dst, Translator tr): dst_{ std::move(dst) }, translator_{ std::move(tr) } { }
 
   template<typename ...Args>
-  void log(Args&& ...args) const { dst_->log( std::forward<Args>(args)... ); }
-
-  template<size_t N, size_t M, typename ...Args>
-  void log(Format::ParsedCompiletime<N,M>&& parsed, Args&& ...args) const
+  void log(Args&& ...args) const
   {
-    const auto translated = translator_->translate( std::move(parsed) );
-    const auto formatted = Field::FormattedString{ Format::apply(translated, args...) };
-    dst_->log( formatted, std::forward<Args>(args)... );
+    Data tmp;
+    tmp.reserve( sizeof...(args) );
+    convertArgs( tmp, std::forward<Args>(args)... );
+    Backend::FieldInfo fi{ Log::Destination::Common::rootElementTag(), std::move(tmp) };
+    dst_->log( std::move(fi) );
   }
 
   void reload() { dst_->reload(); }
   void flush() { dst_->flush(); }
 
 private:
+  using Data = std::vector<Backend::FieldInfo>;
+
+  void convertArgs(Data&) const { }
+
+  template<typename Head, typename ...Tail>
+  void convertArgs(Data& out, Head&& head, Tail&& ...tail) const
+  {
+    using Backend::toFieldInfo;
+    out.push_back( toFieldInfo( std::forward<Head>(head) ) );
+    convertArgs( out, std::forward<Tail>(tail)... );
+  }
+
+  template<size_t N, size_t M, typename ...Tail>
+  void convertArgs(Data& out, Format::ParsedCompiletime<N,M>&& parsed, Tail&& ...tail) const
+  {
+    const auto translated = translator_->translate( std::move(parsed) );
+    auto formatted = Field::FormattedString{ Format::apply(translated, tail...) };
+    convertArgs( out, std::move(formatted), std::forward<Tail>(tail)... );
+  }
+
   mutable Destination dst_;
   const Translator translator_{};
 };