+++ release_notes.md
@@ -6,6 +6,7 @@ short summary of each release, with version and release date.
 work in progress.
 this work is available on "master" branch, but has not yet been released.
 
+  * arguments before `BUT_FORMAT` are now allowed as well (these do not participate in formatting).
   * added `Log::Destination::AutoFlush`, that forwards all logs to a `Sink` and automatically flushes, after each call.
   * added `Log::Destination::BackgroundThread`, that forwards all logs to a `Sink`, in a separate thread.
   * disabled tests for `ThreadPool`, using `boost` backend - GCC's thread sanitizer detects data races in boost...